requirements.txt
Flask
beautifulsoup4
requests
instagrapi
python-dotenv